--- airbyte-cdk/bulk/core/load/src/integrationTest/kotlin/io/airbyte/cdk/load/mock_integration_test/MockDestinationSpecTest.kt
@@ -1,11 +1,0 @@
-/*
- * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.cdk.load.mock_integration_test
-
-import io.airbyte.cdk.load.spec.SpecTest
-import io.airbyte.cdk.load.test.mock.MockDestinationBackend.MOCK_TEST_MICRONAUT_ENVIRONMENT
-
-class MockDestinationSpecTest :
-    SpecTest(additionalMicronautEnvs = listOf(MOCK_TEST_MICRONAUT_ENVIRONMENT))
