+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/command/EnvVarConstants.kt
@@ -22,11 +22,6 @@ data class Property(val micronautProperty: String, val environmentVariable: Stri
 
 object EnvVarConstants {
     val AIRBYTE_EDITION = Property("airbyte.edition", "AIRBYTE_EDITION")
-    val FILE_TRANSFER_ENABLED =
-        Property(
-            "airbyte.destination.core.file-transfer.enabled",
-            "USE_FILE_TRANSFER",
-        )
     val RECORD_BATCH_SIZE =
         Property(
             "airbyte.destination.core.record-batch-size-override",
