+++ airbyte-cdk/bulk/core/load/src/main/resources/application.yaml
@@ -11,9 +11,6 @@ airbyte:
       mappers:
         namespace-mapping-config-path: ${NAMESPACE_MAPPING_CONFIG_PATH:/dest/namespace-mapping.json}
       record-batch-size-override: ${AIRBYTE_DESTINATION_RECORD_BATCH_SIZE_OVERRIDE:null}
-      file-transfer:
-        enabled: ${USE_FILE_TRANSFER:false}
-        staging-path: ${AIRBYTE_STAGING_DIRECTORY:/staging/files}
       resources:
         disk:
           bytes: ${CONNECTOR_STORAGE_LIMIT_BYTES:5368709120} # 5GB
