+++ airbyte-cdk/bulk/core/load/src/testFixtures/kotlin/io/airbyte/cdk/load/component/ConnectorWiringSuite.kt
@@ -167,8 +167,8 @@ interface ConnectorWiringSuite {
             val count = client.countTable(tableName)
             assertEquals(1L, count, "Should have exactly 1 record after write. Got $count records.")
 
-            // 6. Close loader
-            loader.close(hadNonzeroRecords = true, streamFailure = null)
+            // 6. Teardown loader
+            loader.teardown(completedSuccessfully = true)
         } finally {
             // Cleanup
             client.dropTable(tableName)
