+++ airbyte-cdk/bulk/toolkits/load-object-storage/build.gradle
@@ -1,10 +1,10 @@
 def kotlinCoroutinesVersion = "1.10.2"
 
 dependencies {
+    // Needed for Micronaut annotations and logging
     implementation project(':airbyte-cdk:bulk:core:bulk-cdk-core-base')
-    implementation project(':airbyte-cdk:bulk:core:bulk-cdk-core-load')
 
-    // load-object-storage uses pipeline/task infrastructure from legacy-task-loader
+    // load-object-storage is a legacy toolkit - get all dependencies through legacy-task-loader
     api project(':airbyte-cdk:bulk:toolkits:bulk-cdk-toolkit-legacy-task-loader')
 
     // TODO: Separate individual format types from the core format spec
